Api requires an authenticated user

+++ src/Controllers/ApiController.php
@@ -23,6 +23,8 @@ class ApiController extends Controller
      */
     public function index(): AnonymousResourceCollection
     {
+        abort_unless(Auth::check(), 401);
+
         return SocialAccountResource::collection(Auth::user()->socialAccounts);
     }
 
@@ -35,6 +37,8 @@ class ApiController extends Controller
      */
     public function show(SocialAccount $socialAccount): SocialAccountResource
     {
+        abort_unless(Auth::check(), 401);
+
         return new SocialAccountResource($socialAccount);
     }
 
@@ -47,6 +51,8 @@ class ApiController extends Controller
      */
     public function destroy(SocialAccount $socialAccount): JsonResponse
     {
+        abort_unless(Auth::check(), 401);
+
         if ($socialAccount->delete()) {
             return response()->json([
                 'message' => 'Deleted',
